Projeto finalizado

+++ application/views/pages/games.php
@@ -13,8 +13,8 @@
 					<th>#</th>
 					<th>Name</th>
 					<th>Price</th>
+					<th>Category</th>
 					<th>Developer</th>
-					<th>realease_date</th>
                     <th>Actions</th>
 				</tr>
 			</thead>
@@ -24,17 +24,17 @@
                         <td><?=$game['id'] ?></td>
                         <td><?=$game['name'] ?></td>
                         <td><?=$game['price'] ?></td>
+						<td><?=$game['category'] ?></td>
                         <td><?=$game['developer'] ?></td>
-                        <td><?=$game['release_date'] ?></td>
                         <td>
-							<a href="<?= base_url() ?>games/edit/<?= $game['id'] ?>"
-							class="btn-sm btn-warning">
-							<i class="fas fa-pencil-alt"></i></a>
-
-							<a href="<?= base_url() ?>games/delete/<?= $game['id'] ?>"
-							class="btn-sm btn-danger" onclick="return confirmDelete('<?= $game['name'] ?>');">
-							<i class="fas fa-trash-alt"></i></a>
-							
+							<?php if($_SESSION["user_data"]["user_id"] === $game["user_id"]) : ?>
+								<a href="<?= base_url() ?>games/edit/<?= $game["id"] ?>" class="btn btn-warning btn-sm"><i class="fas fa-pencil-alt"></i></a>
+								<a href="<?= base_url() ?>games/delete/<?= $game['id'] ?>"
+							class="btn btn-sm btn-danger" onclick="return confirmDelete('<?= $game['name'] ?>');">
+							<i class="fas fa-trash-alt"></i></a>							<?php else : ?>
+								<button disabled type="button" class="btn btn-warning btn-sm"><i class="fas fa-pencil-alt"></i></a>
+								<button disabled type="button" class="btn btn-danger btn-sm"><i class="fas fa-trash-alt"></i></a>
+							<?php endif; ?>
 						</td>
                     </tr>
                 <?php endforeach ?>
